+++ changelog.md
@@ -1,4 +1,5 @@
 # 3.3.1
 
 - Removed placeholder from input icon value. [#130](https://github.com/shahabyazdi/react-multi-date-picker/issues/130)
+- mobileButtons added to props
 - Dependencies updated. (React 18.2.0 and ...)